requirements.txt
google-genai>=1.0.0
pandas>=1.5.0
sqlalchemy>=2.0.0
pypdf>=4.0.0
typer>=0.9.0
rich>=13.0.0
tqdm>=4.64.0
numpy>=1.20.0
